Allow XenStore reads and writes independent of watch status.
libxenbackend delegates certain responsibilities (i.e. reacting to
XenStore watch events) to the caller, which must route events back
to libxenbackend.

Prior to this commit, if these events occur during libxenbackend
initialization, it is possible for a triggered watch to block the
XenStore socket, resulting in deadlock on the next XenStore transaction
unless the caller has a second thread monitoring the XenStore fd.

This commit separates the watching and general XenStore access
into two separate sockets, ensuring that the general access socket
will never be blocked, and avoiding deadlock.